fix: Long으로 파싱되지 않는 문제 해결

+++ api/src/main/kotlin/com/few/api/security/token/TokenResolver.kt
@@ -40,7 +40,8 @@ class TokenResolver(
                 .build()
                 .parseClaimsJws(token)
                 .body
-                .get(MEMBER_ID_CLAIM_KEY, Long::class.java)
+                .get(MEMBER_ID_CLAIM_KEY, Integer::class.java)
+                .toLong()
         } catch (e: Exception) {
             log.warn { "${"Failed to get memberId. token: {}"} $token" }
             return null
